Address PR comments: remove console statements, rewrite getCategories; split ReportLink into own component and remove from Error
