+++ src/components/icon/icons.jsx
@@ -58,7 +58,6 @@ const Icons = ({ icon }) => {
         />
       );
     default:
-      console.warn("Unknown icon type:", icon);
       return null;
   }
 };
