+++ src/pages/categories/index.jsx
@@ -5,7 +5,6 @@ import Loader from "~/components/loader";
 import Error from "~/components/error";
 import { H1 } from "~/components/typography";
 import api, { useAPI } from "~/core/api";
-import theme from "~/core/theme";
 import CategoryCard from "./category-card";
 
 const StyledLogo = styled(Logo)({
@@ -25,37 +24,19 @@ const IntroText = styled(H1)({
 });
 
 // Transforms the data into usable objects and filter out duplicates
-const getCategories = data => {
-  const categoriesBySlug = {};
-  data.forEach(([title, slug, description, icon, color]) => {
-    const existingCategory = categoriesBySlug[slug];
-    const category = {
-      color,
-      description,
-      icon,
-      slug,
-      title,
-    };
-    if (theme.colors[color] === undefined) {
-      console.warn(`Unknown icon color: ${color}`);
-    }
-    if (!existingCategory) {
-      categoriesBySlug[slug] = category;
-    } else {
-      // Check for duplicate values; warn if we see them
-      ["title", "description", "icon", "color"].forEach(fieldName => {
-        const existingVal = existingCategory[fieldName];
-        const val = category[fieldName];
-        if (existingVal !== val) {
-          console.warn(
-            `Entries with category slug ${slug} have different ${fieldName} values: ${existingVal} and ${val}`
-          );
-        }
-      });
-    }
-  });
-  return Object.values(categoriesBySlug);
-};
+const getCategories = data =>
+  Object.values(
+    data.reduce((categories, [title, slug, description, icon, color]) => {
+      categories[slug] = categories[slug] || {
+        color,
+        description,
+        icon,
+        slug,
+        title,
+      };
+      return categories;
+    }, {})
+  );
 
 const Categories = () => {
   const { loading, errorMessage, data } = useAPI(api.getIndex);
